feat: Add random number seed option to main.go
The code changes add a new command-line flag `seed` to the `main.go` file, allowing the user to specify a seed for random number generation. If a seed is provided, the `rand.Seed()` function is called with the provided seed value. This change enhances the flexibility and reproducibility of the simulation.